Modularize developer integration settings
Separate CLI and MCP setup, hosted API controls, local API and webhooks, clipboard handling, and API key rows behind the existing developer settings facade without changing behavior.
